Create and Implement Protocol
Create a Display protocol, and implement it for the Action and
Character modules using info/1. Remove the Enum mapping since the
data types handled by the protocol are being passed directly and the
protocol will display.

Also implement Elixir's String.Chars protocol on the Character and
Action modules, returning as a string the name or label respectively