Implement Battling
Add an enemy trigger that will randomly select an enemy and begin a
battle. In the battle, characters will swap attacks recursively until
one has zero hit points. If "you" win, you can move to the next room.

The end of the game is finalized - if the next room has an enemy, the
battle sequence continues. If the next room is the exit, "you" win the
game.